grid-cell: implement star emblems
Nautilus does not display anything if the user selects "Star" as caption for
for the grid view.

This is because the star caption quark is not handled by
nautilus_file_get_string_attribute_q, which causes us to set NULL for
the caption label.

As a more definitive solution, implement the display of star emblems
for the grid view instead of relying on captions.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2842