file-operations: make use of skip_files in compress operations
Nautilus fails to create an archive if one of the folders selected
cannot be read, despite selecting "Skip".

This is because the compress operation does not use skip_files, which
was populated by scan_sources.

Remove source_files nodes if they should be skipped before passing the
linked list to AutoarCompressor.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2928