mime-action: Use GtkFileLauncher
While sandboxed, we should use a portal to open files with another app.

But the portal is not getting used for archive file types or any of
their subtypes. Instead, nautilus itself tries to handle them. Such is
the case for Open Document file types, which are subtypes of ZIP.

This happens because GAppInfo uses the portal only as a fallback for
when it doesn't find any handler in the sandbox. But it finds nautilus
itself inside of the sandbox, which claims to handle archives, so it
picks it.

Instead of using GAppInfo, use the new GtkFileInfo API which uses the
portal directly. Also, remove the now unused hack, because GtkFileInfo
does everything we need while running as a flatpak.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2841