search-engine: Do not check "available" sources
By default, the filesystem miner only indexes folders within the
user homedir which are always considered available. Getting anything
else here means the user fiddled directly with dconf settings not
exposed anywhere and indexed mountpoints with intermitent availability.

Accounting for the remote possibility of this being ever false
means we have to jump a few hops to reach from content to file to its
data source, which may become a noticeable bottleneck if the FTS search
terms match a lot of files.

While indexing of removable devices has plenty of room for improvement
(starting with the tracker-miners side) and should be a prime feature,
it remains a niche option. Do not optimize for the unlikely and
optimize for large result sets.