search-engine-simple: Delay starting IO
While the user types a search query, we restart our search engines for
every key stroke. This is desirable for quick results.

But unlike other engines, starting the simple engine multiple times in
rapid succession can put some stress on system resources.
Depending on the hardware and windowing system, this may result in user
perception of poor performance.

Previously, we've dealt with this by setting the delay directly in the
search entry. But that was a bad solution because it draged down the
responsiveness of other search engines.

Instead, let's save resources by starting the simple engine IO thread
half a second later, if not cancelled/restated before that.

Bump glib minimum version to use "once" source helpers.