search-engine: Use recent engine only for shell search
The recent engine has been introduced mainly for systems without
tracker, which used to be the case with Ubuntu.

But nowadays we don't really support not running tracker.

So the only value this engine might bring is for shell search
provider, where we don't want to run the IO-intense simple engine
as fallback. In that case, the recent engine may find files that
are not indexed by tracker.

In all other cases, these files are going to be found by the simple
engine anyway. Not only is the recent engine doing futile IO on
every keystroke, it's another engine which we need to cancel and
wait to finishe before restarting all engines.

Also add some missing debug messages which helped diagnose this.