window-slot: Don't add search to history
We've been treating search as a location that can be "opened" and
even navigated back or forward to.

Yet, the results of navigating back or forward to a search directory
are inconsistent and even unpredictable.

Furthermore, from a design point of view, search should be treated as
filtering the current location, not as navigating to another location.

If that were not enough, treating searching as opening a location
triggers a cascade of code which ends up reloading it twice. This
means search engines are restarted twice: START-STOP-START-STOP-START.
Starting engines 3 times for every keystroke is not only wasteful,
it multiplies any performance bottlenecks when starting and stopping
search engines.

In short: an unpredictable feature that reflects implementation
details rather than UX design and amplifies search performance issues.

In order to perform a search without causing a navigation event, let's
keep the slot location unchanged. This means the slot location and
the view location get out of sync, so we need extra logic to update
the slot title accordingly.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2686