search-engine-tracker: Don't block main thread on start
We check whether to recurse right when we start the tracker engine.

But when recursive search is local-only (the default option), this
means we need to check whether the file is local. Currently, we do
this using sync GIO API. But the start method happens in the main
thread, and doing sync IO in the main thread is going to block
everything, causing serious performance issues for some setups.

Instead, it can just use the filesystem info that's already cached
by the NautilusFile instance for the searched directory.

This means the only consumer of search-engine-private.h becomes
NautilusSearchEngineSimple, which safely performs the check in
another thread, so we can move the whole logic there to simplify.

This effectively reverts [0], but also reapplies the bugfix
from [1] in form given by [2].

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2967

[0] de36e000f4bf38fdf4698126bae0dcc49bd9ffbc
[1] 62ba6d83df162b149785ed10a13d242ee1a13dc1
[2] f0759d6fe9bb405c3e9a5a5cf2e9509514f4eb6f