files-view: Appear to filter while searching
When loading search results, we clear all items in view, even before
we get the new list of results.

This results in the view going blank for a moment. Now that search
for results in the current folder is no longer delayed, this blank
flashing happens on every character entered, which is annoying. Also,
it further contributes to the perception that searching is slow.

Instead, delay the clearing until the next batch of search results is
ready to load. This way, clearing the view and refilling it happen
together, resulting in an illusion of the same view getting filtered.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2972