style: Add real neutral spacing between items
We've been adding margins to items to keep them appart.

But the rubberbanding algorith doesn't account for these margins, so,
e.g., when drawing a selection rectangle in the list view by starting
from the space between rows, the closest row is included in the
selection even if the selection extends in the opposite direction.

Now that GTK has gained support for the border-spacing CSS property in
GtkListBase subclasses, which the rubberbanding algorithm accounts for,
we can finally fix this bug in list view. For grid view this is not
a complete fix, because items have variable sizes. But at least there
is a safe space.

Resolves a large part of https://gitlab.gnome.org/GNOME/nautilus/-/issues/2974