general: Add detailed date and time format option
The default time and date formats in Files are intended to be human
friendly and to avoid information overload. This works for most cases
and is an appropriate default. However, there are cases where people
need to be able to see a complete date and time for their files. Let's
add a user preference to show a numerical date and time format.

Resolves: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2874