preferences-window: Don't keep GtkBuilder beyond building
The signals keep a reference to the builder, which keeps the window and
its widgets alive indefinitely, even after the window is destroyed,
resulting in a leak.

We are using GtkBuilder solely as a way to access pointers to the
captions combo rows. That job could just as well be done by an array.