window: Only allow tab dnd in wayland
There's a bug in x11 GtkDropTarget's where the action on the GdkDrop
isn't set correctly.  We have a workaround for this in
nautilus-list-base (where we get the action from the GdkDrag).  But
in AdwTabBar, we don't get the GtkDropTarget (or the GdkDrag/Drop),
so the hack isn't possible.

DnD won't work in x11, so disable it, rather than giving false promises.