view-item: Don't propagate view icon size through the model
Currently, icon size is by each subclass (list and grid views), which
call a private method to set the icon size on every item in the model.
Name and Grid cells get the icon size from the model item and listen to
its change notifications. In short, this is the information flow:

    NautilusListView -> NautilusViewItem:icon-size -> NautilusNameCell
    NautilusGridView -> NautilusViewItem:icon-size -> NautilusGridCell

This is conceptually wrong, because icon size is an attribute of the
view, not of each individual item, and the view model shouldn't be
used to keep view state. Also, this means every item in the model
is updated even if it's not currently bound to any cell.

Thankfully, we have abstract classes for both views and cells, so we
can implement the following information flow with bindings:

    NautilusListView:icon-size -> NautilusNameCell:icon-size
    NautilusGridView:icon-size -> NautilusGridCell:icon-size

With this new workflow, NautilusViewItem:icon-size is no longer used,
so it's dropped. This change prepares the view model to be loaded by
both views simultaneiously, for fast view switching without reloading.