window-slot: Make "back" leave search first
Quoting Allan:

On the one hand, search is treated as a filter on the current location,
so conceptually I would expect it to be independent of navigation
controls. On the other hand, I'm not sure if there's anything better
for forward and back to do in this context. Having them actually take
you to another location does seem like it would be surprising.