gtkplacesview: Allow smaller width
Currently, when the window is resized to its minimum width before
opening "Other Locations", the window is made larger to fit the
"Connect to Server" bar.

This behavior is generally undesirable and it's unusable on mobile.

In order to resolve this issue, implement some changes from the
mockups of the future Network location which uses a simplified bar
without the "Connect to Server" label:
https://gitlab.gnome.org/GNOME/nautilus/-/issues/2785

A few more changes were necessary:

 * GtkActionBar wouldn't allocate the whole width to start-packed
   children, so it's replaced with a toolbox-styled GtkBox plus a
   GtkSeparator.
 * GtkEntry is given a smaller minimum with. It's also given a
   maximum width as per mockups-
 * GtkButton is allowed to shrink, because some translations of
   "Connect" might be too long to fit.

Part of https://gitlab.gnome.org/GNOME/nautilus/-/issues/2448