grid-view, list-view: don't sort directories before files in search results
As described in #2968, directories are currently sorted before files in search results
if this option is enabled in preferences, leading to less relevant results being shown
first if they are directories. Intended behaviour is for this preference to be ignored
when searching. This bug was originally fixed in 9ec56a2d48ddf5db18757e680deb2421c15189da,
it is caused by the directories_first check being made at init, which is before the
location has been set.

This commit moves the directories_first check to real_begin_loading in the grid view
and list view classes. This causes the directories_first preference to be ignored when
presenting search results, as expected.

Closes #2968