grid-cell: Show location as a caption in search & friends
Search results from subfolders are the main challenge with using grid
view for search results. For these results, we need to show the path
in order to differentiate between otherwise identical results.

Show the path as a caption for search and, for consistency, also for
the other special locations where list view shows a path.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2969