window: Make the main headerbar flat
The shadow under the headerbar doesn't look good in combination with
the border under the tab bar. Also, it becomes redundant with the
shadow under the column view headers when scrolled.

Make it flat, integrate the tab bar into the top toolbar, and apply
the "view" style to the whole content page.