sidebar: Hide format for unmounted MTP/GPhoto2 volumes
Currently, the Format action is offered also for the MTP volumes when
they are unmounted. This is a regression from the commit ab1d3019. This
is because the `uri` variable is `NULL` for unmounted volumes. This
amoung other causes critical warnings when opening popover menu for
unmounted volumes. Let's check that the activation root is native file
instead to avoid the mention issue. This will also correctly hide the
Format action also for GPhoto2 devices that are mentioned in the original
commit, but unhandled currently.