file-operations: Do png conversion for image saving on worker thread
We are calling gdk_texture_save_to_png_bytes() on the main thread.
This was originally done in order to avoid calling a gdk function
on another thread.  GdkTexture's are immutable, however, and all the
work done retrieving it is on the main thread.  In order to prevent
blocking the main thread, let's move this call off the main thread.