files-view: Update NautilusPreviewer on slot changed
Nautilus sends updates to previewer when there's changes to the view's
selection.  This works in all cases, except context switches where
the view's selection hasn't actually changed.  When we are switching
tabs, although the selection for the view hasn't changed, the selection
in terms of how the user perceives it *has* changed.

I'd argue that the logic is slightly more complicated when switching
between two  windows. For example two windows could be side by side,
in which case the user won't want to switch the previewer, unless
specifically selecting a new file in the other window).

Send previewer updates when switching tabs (slots)