previewer: Respond to selection events on actual window
SelectionEvent's (pushing arrow keys from the previewer window), are
emitted to all subscribers of the signal.  When there's one instance
of Nautilus with a single window, this is not a problem.  But when
there's multiple instances (i.e. Nautilus nightly and Nautilus, or
another subscriber to NautilusPreviewer2), the SelectionEvents
get consumed by all subscribers even though they probably don't
expect it.

This issue can be solved with the new ParentHandle property of
NautilusPreviewer2 which returns the parent window of the previewer
window.  Only respond to the SelectionEvent if the ParentHandle
exists in this instance of Nautilus.