file: Translate time formats according to time locale
We have different date and time formats for different cases, and
they all need to be translated.

When the user picks a locale for date-time formatting that's from a
language different from the language chosen for the UI text, we may
get a broken date string, with the format from one language but
pieces of text from another language.

Fix this by forcing gettext to translate these format strings
using the same language as LC_TIME.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3004