list-view: Simplify sort action management
We've been activating the view.sort action when the columnview sorter
changes as a result of the user clicking the header to sort by that
column.

However, the action itself changes the sorter, which causes another
emission of the ::changed signal. In order to break the infinite
reentrancy loop, we use a boolean flag.

But there is a simpler solution: updating the action state without
activating the action, whenever the sorter changes. Let's do that.
And also properly document this code.