file: Import SortType to attribute mapping
Keeping this mapping in the view code forces us to use static
attribute strings, which risks getting out of sync with the
ones used by the file code.

Furthermore, the file code already does the reverse mapping.

This is going to help with moving non-view implementation code
out of nautilus-list-base.c