files-view: Don't save sort metadata on visit
We save sort options when the state of the view.sort action changes.

But this state may be changed in cases where the user hasn't changed
the sort options. Such is the case when the folder is loaded and we
read preferences and metadata to update the state of the sort action.

As a result, visiting a folder for the first time causes the default
sort setting to be saved as metadata and remembered in future visits.
This makes the setting moot except for never before visited folders.
This also causes us to save metadata unnecessarily.

Instead, sort metadata should be saved only when the user explicitly
changes the sort order form the default on each folder. To enforce
this, block our handler of state changes when updating from settings
and metadata.

Resolves a partial regression https://gitlab.gnome.org/GNOME/nautilus/-/issues/2931#note_1784938