starred-directory: Don't emit ::files-added when [re]loading
On force reload, we clear the list of starred files and repopulate it.

But emitting NautilusDirectory::files-added when repopulating means any
view displaying this directory, other than the one being reloaded, is
going to react to this signal by adding to the view files that were
already in view, duplicating them in the view UI.

Also populate the list on init() using the same codepath, while nobody
is listening to our signals yet. In that case it's safe, but useless.

As such, we should stop emiting the signal.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2998