tag-manager: Remove duplicate emissions of "starred-changed"
We notify the changes we make change the database, then we get the
signal from the database about these changes, and we notify them again.

The first emission is untimely, because it leaves us in an inconsistent
state: the `starred_file_uris` hash table hasn't been updated yet, and
any clients calling `nautilus_tag_manager_file_is_starred()` to assess
the result of the notified change get the wrong answer.

So, let's stop emitting "starred-changed" before we get the database
signal about the changes. This also resolves a leak of a files list
from update_moved_uris_callback().