libnautilus-extension: Reformat documentation
Our current documentation is really poor.  Almost all of the short
descriptions are missing (or poorly formatted).  Most of the links
are incorrectly formatted. Some methods/properties/signals are
missing documentation altogether.  Since I am changing the
documentation so extensively, move it to the .c files
which is the correct location (so the links can be properly created
to the gitlab source).