exp-app: Move NautilusColumnProvider extension to extension manager
NautilusColumnProvider only needs to be loaded once because it provides
static metadata of the column.  Call nautilus_column_provider_get_columns
when the extension manager loads and set it as a property (using the
four attributes of a column).  Since it's a property, nautilus can
take advantage of the cached property call.