list-base: Stop handling ::begin-loading
The plan is for NautilusListBase implementations to become children
of NautilusFilesView, not a subclass.[0] So we need to keep removing
our vfunc subclass overrides.

Let's take another step in that direction: stop using ::begin-loading,
instead expect NautilusFilesView to ask its (future) child to set
itself up for the new directory.

(The new_directory argument is not used yet, but will be important for
the next commit.)

[0] https://gitlab.gnome.org/GNOME/nautilus/-/issues/3042