file: separate uid and gid existance property
Store information on whether UID/GID is set in separate boolean and
add checks that it is set before reading the value. This removes the
need to rely on the special -1 value that prevented storing the IDs as
their respective actual types, uid_t and gid_t.

This fixes un<>signed comparison warnings.