query-editor: Make :location write-only
This property is meant to be used only as a "socket" for the slot to set
the current location, for use in new queries. This is used internally
and, as such, it makes no sense to expose a getter.

The location being searched should be obtained from the query itself,
not the query editor. For that we have a get_query(), and the query
parameter in the ::changed signal.