files-view: Suggest global search if no matches
If you type the right search term while in the wrong location, you need
to abort search, move to a correct location, then type again.

This is both frustrating and, sometimes, hard if you don't know the
right location. You may know what you are looking for but not where it
is located.

Let's offer to apply the same search terms into a global search query,
with a suggested-action pill button. The statuspage style must be
adapted not to fade the new button.

Part of https://gitlab.gnome.org/GNOME/nautilus/-/issues/3013