files-view: Offer to make search global
If you have seen every result (e.g. scrolled to the bottom) and didn't
find the file you need, you end up in the same situation described in
the previous commit when no result is found.

But in this case we don't want to make it a strong suggestion.

Have a handy but unobstrusive button to "globalize" search if the
results don't fill the whole view, or if the view has been scrolled
till the end.

Part of https://gitlab.gnome.org/GNOME/nautilus/-/issues/3013