sidebar: Don't navigate into trash on drag-and-drop
When a user drags a file onto a location in the sidebar, the location
will open after a short duration, to allow users to drop the file in a
subdirectory of the location. However this doesn't make sense for trash
because you can't drop a file in a subdirectory in trash.

Fixes #2831.