file: Actually ensure alphabetical order in search
Files from the current folder sort in reverse alphabetical order
in list view, but not in grid view.

This is because grid view explicitly asks for reverse order, while
list view asks for direct order and lets the GtkColumnViewSorter
handle the reversal.

Instead of assuming we always compute the reverse order ourselves,
explicitly reverse the alphabetical order while sorting by search
relevance. This works both in grid and list view.