window: Remove duplicate progress indicator
When adapting to small widths, some widgets are "moved" to from the
top toolbar to the bottom toolbar at small widths.

But the progress indicator has moved to the sidebar, so this is no
longer necessary. Also, we were not really moving the widget anyway,
we had a duplicate widget, showing only one at a time. This is the
cause of duplicate toasts.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3056