list-base: Allow popover focus
We intercept focus tabbing before passing it to children in order to
emulate the tab behavior that would be introduced in gtk 4.12. But
this was really only meant to affect the content.

Instead it's affecting popovers attached to the view as its children.
So, adjust the logic to check the ancestor is the content widget.

Closes https://gitlab.gnome.org/GNOME/nautilus/-/issues/3052