window: Adjust sidebar width
The sidebar is a bit too wide at any window size.  Yet
the sidebar would be too small in a large window if we just
reduced max-sidebar-width to a much lower amount. Let's
bump down max-sidebar-width a little, but also decrease
sidebar-width-fraction to keep the sidebar smaller for
smaller windows per discussion IRC.