files-view: Don't check for non-ready files
Back in 2001, after the nautilus 1.0.4 release, many changes were made
to directory async loading[0], including support for desktop entry
files of types Link and Directory[1,2].

This was a user-visible change: instead of all icons appearing at once
when the folder is fully loaded, the user would see icons being added
incrementally as the folder loads.[3]

This had 2 consequences:

  - icons "dancing around" as new icons changed grid sizes,[4]
  - performance problems, with slower loading.[5]

The first bug was inherent to the canvas widget and wouldn't be
erradicated until the canvas itself was decommissioned[6].

The second bug was a blocker for the 1.0.5 release. Developers tried[7]
to tweak the update interval (hysteresis)[8] or the size of chunks[9],
but doing that only exposed more performance problems.

Then Darin Adler diagnosed the source of the slowness. It was the async
job looking for .desktop link files to read them and load the custom
attributes (display name, custom icon) declared inside the file.[10]

On one hand, it was undesirable to display .desktop files before they
were read, because their name and icon would change. On the other hand,
reading files synchronously would risk freezing the app in some cases
(e.g. while loading a remote directory). Trying to avoid both problems
lead to the slow loading problem.[11]

The solution to work around this bug for the nautilus 1.0.5 release was
the introduction of a triage step in the displaying of pending files:
non-ready files (such as .desktop files waiting for their attributes to
be read async) would be moved into a separate "waiting room", not to
block the rest of the queue.[12]

This succeeded in addressing the regression and 1.0.5 was released.

Fast forward to 2023, .desktop links support is long gone[13], but the
very same "waiting room" logic remains. We no longer have to read any
file to know their "display name", but new thumbnails generation still
happens asynchronously from view loading. This means we still see files
being added to the screen as their thumbnails are created, pushing other
files around.[14]

Unlike 22 years ago, multi-core CPUs are common place, powering recent
optimizations which speedup new thumbnail generation.[15] Even if the
thumbnail is not immediately available, we can show a placeholder icon.
And the icon grid has a fixed width, so it's not affected by thumbnails
aspect ratio.

Therefore, there is really no reason to hold off adding files to the
view until their icons/thumbnails are ready. So, drop the waiting room
to make view loading complete ealier and feel smoother.

(Diagnosis and patch by Corey Berla, storytelling by António Fernandes)

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2826

[0] https://gitlab.gnome.org/GNOME/nautilus/-/blob/829732a43eb6686c35fcb8467f6c754be8a27069/ChangeLog
[1] Still mentioned, but not explained, in the spec: https://specifications.freedesktop.org/desktop-entry-spec/latest/
[2] Example of a *.directory file: https://gitlab.gnome.org/GNOME/nautilus/-/blob/5d691d775097c9d73f4e544ec36c58bec1c05b4c/data/starthere.directory.in#L4
[3] https://mail.gnome.org/archives/nautilus-list/2001-October/msg00182.html
[4] https://mail.gnome.org/archives/nautilus-list/2001-October/msg00186.html
[5] https://mail.gnome.org/archives/nautilus-list/2001-October/msg00188.html
[6] https://gitlab.gnome.org/GNOME/nautilus/-/commit/2d1deaac2dd12b0ba16446bfbf3498b266e60338
[7] https://mail.gnome.org/archives/nautilus-list/2001-October/msg00200.html
[8] Hysteresis was later dropped in https://gitlab.gnome.org/GNOME/nautilus/-/commit/6a263ae188943ac18f476b80e3b836c6c24b3c35
[9] Chunk processing was later dropped in https://gitlab.gnome.org/GNOME/nautilus/-/commit/b9bbfaa39b1af3ce90e99762a8cb613a5db3af3d
[10] https://mail.gnome.org/archives/nautilus-list/2001-October/msg00343.html
[11] https://mail.gnome.org/archives/nautilus-list/2001-October/msg00333.html
[12] https://gitlab.gnome.org/GNOME/nautilus/-/commit/5d691d775097c9d73f4e544ec36c58bec1c05b4c
[13] https://gitlab.gnome.org/GNOME/nautilus/-/commit/b1cb977812054a45bb10a4b297871a25ad3262d9
[14] Closes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2826
[15] https://gitlab.gnome.org/GNOME/nautilus/-/commit/78644dcb6f1dfdc8db290cca58575949d7517fa8