list-base: Set selected files more efficiently
We call `nautilus_view_model_get_index()` once for each item to select,
which is means we are repeatedly iterating the model once for every
item to select.

The number of items to select is usually smaller than, and at most
equal to, the number of items in the model.

So, let's instead iterate the model once, looking for items in the
selection subset. Also, use g_list_find() and remove files as they
match, to make it even faster.