files-view: Stop delaying pending selection reveal
There are many situations where we open a folder to highlight an item:

 - the "Open Item Location" action in Recent, Starred and Search,
 - the --select command line option,
 - the "Show in Folder" action in the web browser's list of downloads,
 - the "Open Containing Folder" in file viewer apps,
 - when going "up" to the parent folder

In such cases, the item to highlight is called "pending selection".

When we the view scrolls, it's necessary to adjust the scroll position
to reveal the pending selection. But this adjustment is delayed until
after the view is loaded and rendered into the screen. This results
in slow and ugly user experience: first the top of the folder is shown,
then, after a variable ammount of time, the view suddenly jumps to
reveal the pending selection. The sudden jump may even lead to data
loss if the user clicks a file and doesn't notice the desired file
was moved under the pointer, as reported in [0].

This delay was added 17 years ago[1] to wait for GtkTreeView nodes to
be allocated before trying to scroll to a selected node.

GtkTreeView was dropped as part of the GTK4 port, but the new views had
the same problem, except with real child widgets which needed to be
allocated before we could scroll into their coordinates. And allocation
happens only when the frame clock ticks.

But now that the new views have a proper scroll API[2], which keeps
track of list item positions, it doesn't matter whether the item widget
is allocated or not. So we can simply call that API immediately.

Combined with faster view loading[3], the result is that the pending
selection is revealed right away, with the directory being shown
already in the desired scroll position.

[0] Closes https://gitlab.gnome.org/GNOME/nautilus/-/issues/3014
[1] 53103444cc1cae054d922857391d1f5d19ea58ad
[2] 710fa773d67d64208fd24789264e533c2b492049
[3] a7299fe3fc5179f63ae4fc0f41371b5c9f6bb99f