view-model: Advise against calling get_index() recursively
`nautilus_view_model_get_index()` is `O(N)`: it iterates the whole
model linearly to find the item. It should never be used in a loop
because that makes the whole thing `O(N^2)`.

The previous commits have resolved all such occurences.

In order to prevent new occurences, document the function with a
warning about this.