window: Change sidebar title
The new split view design has a full-height place sidebar with its 
own headerbar and, as such, its own title. The design mockups used
the old "Places" title for this. 

The name "Places" has been used prominently in the GNOME 2 days, 
for a panel menu holding roughly the same list of locations present
in our sidebar nowadays.

However, it's been many years since this word has disappeared UI, 
so it may not resonate the same with newer users. 

The first two entries (Recent and Starred) are not even actual
places. Instead they refer to files. Also, other apps with the
new full-height sidebar designs use the application name as 
title of the sidebar.

So, revisiting the design, we replace "Places" with "Files".