list-base: Drop useless fallback
When sort order changes, it's saved as metadata for the current folder.

If the new sort order is NULL, we save the default sort order instead.
This is pointless (the passed target string is not expected to be NULL)
and wrong (the default sort may change, but metadata isn't updated).

So, just clear the metadata is such unexpected case, in order to have
the default sort order from settings apply.