list-base: Stop jumping to top on Menu key
Pressing the Menu key while the view is scrolled down causes the view
to jump up, which shouldn't happen.

This happens because of a faulty reimplementation of an edge case
handling in the GTK4 port. It's utterly broken and it's my fault.[0,1]

  - The index from the selection filter model is wrongly applied
    to scroll the view (as if it was the index of the unfiltered
    model);
  - If the `for` loop runs until the end, the `i` is not the
    index of the last item: it's the length of the filter model.
  - For list view, the focus child is the inner list view, not
    the cell's parent, so this logic would work only in grid.

The first two points are fixable using GtkBitset instead of
GtkSelectionFilterModel, but the last point means this would
still not work under list view.

The proper solution requires a way to query the focus position,
which doesn't exist (yet? [2]).

Let's give up on the smart handling of corner cases and just get
the common case right. Even in the corner cases, this simple
solution is not too bad.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/3095

[0] 809049a3f6526d5f126d8d4e6918ad483b41baa4
[1] b28e2d545dca3177cf91b3c2bc1855abace719f0
[2] https://gitlab.gnome.org/GNOME/gtk/-/issues/2891