rename-file-popover: create ui via template
Construct the rename popover UI from the .ui file via a template,
instead of using GtkBuilder. As GtkPopover is now the parent class.

Amendment by António Fernandes <antoniof@gnome.org>:
The popover is a child of the view and, as such, the view should be
managing its lifecycle, including the parenting/unparenting. For
that reason, the view is no longer passed to the constructor.