new-folder-dialog: create ui via template
Construct the new folder dialog UI from the .ui file via a template,
instead of using GtkBuilder. As AdwWindow is now the parent class,
replace the NautilusFileNameWidgetController inheritance with template
bindings.

Instead of the previously inherited "name-accepted" and "cancelled"
signals, use a callback. This allows simplifying the dialog's intefrace
and files-view's code.